payment and inventory done

+++ backend/payment-service/src/main/java/com/novamart/payment_service/service/PaymentService.java
@@ -1,5 +1,6 @@
 package com.novamart.payment_service.service;
 
+import com.novamart.payment_service.client.UserClient;
 import com.novamart.payment_service.dto.PaymentRequest;
 import com.novamart.payment_service.model.Payment;
 import com.novamart.payment_service.respository.PaymentRepository;
@@ -14,8 +15,12 @@ import java.util.UUID;
 public class PaymentService {
     private final PaymentRepository paymentRepository;
     private final RefundService refundService;
+    private final UserClient userClient;
 
     public void savePayment(PaymentRequest paymentRequest) {
+        if (!userClient.authenticateUser(paymentRequest.userId(), "accountType", "CUSTOMER")) {
+            throw new RuntimeException("User not authenticated");
+        }
         Payment payment = new Payment();
         payment.setPaymentId(UUID.randomUUID().toString());
         payment.setOrderId(paymentRequest.orderId());
@@ -30,12 +35,18 @@ public class PaymentService {
         paymentRepository.save(payment);
     }
 
-    public Payment getPaymentByOrderId(String orderId) {
+    public Payment getPaymentByOrderId(String orderId, String userId) {
+        if (!userClient.authenticateUser(userId, "role", "VIEW")) {
+            throw new RuntimeException("User not authenticated");
+        }
         return paymentRepository.findByOrderId(orderId);
     }
 
     @Transactional
-    public void deleteAllPayments() {
+    public void deleteAllPayments(String userId) {
+        if (!userClient.authenticateUser(userId, "role", "ADMIN")) {
+            throw new RuntimeException("User not authenticated");
+        }
         refundService.deleteAllRefunds();
         paymentRepository.deleteAll();
     }
